requirements.txt
streamlit==1.29.0
google-generativeai==0.3.2
python-dotenv==1.0.0
pandas==2.1.3
pydantic==2.5.0
typing-extensions==4.8.0
streamlit-chat==0.1.1
requests==2.31.0
